pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fabric>=3.2.2",
    "requests>=2.32.5",
    "setproctitle>=1.3.7",
    "websocket-client>=1.9.0",
]
